Decouple CLI arg parsing/execution for NODE_ENV initialization
