+++ packages/react-router-dev/cli/index.ts
@@ -1,12 +1,16 @@
 #!/usr/bin/env node
-import { run } from "./run";
+import { parseArgs } from "./parse";
 
-run().then(
-  () => {
-    process.exit(0);
-  },
-  (error: unknown) => {
-    if (error) console.error(error);
-    process.exit(1);
-  }
-);
+let { input, flags, command } = parseArgs();
+
+import("./run").then(({ run }) => {
+  run(input, flags, command).then(
+    () => {
+      process.exit(0);
+    },
+    (error: unknown) => {
+      if (error) console.error(error);
+      process.exit(1);
+    }
+  );
+});
